test(logic): split army_integration_test under 400-line gate
CI repo.logic_test_file_size failed on army_integration_test.dart (430 lines).
Move combine/split coverage to army_integration_combine_split_test.dart.

Refs #2316
